fix a bunch of nasty bugs
The "escape" character should really act as if it didn't exist at all,
aka, whether the scanner inserts a concat character or not depends on
whether it should have for the last character before the escape.

Parentheses got pushed to the postfix stack when there's a beam. That's
because I forgot to exclude that case.

The program now returns true if there's a match anywhere in the string.
Before, it returns false if the start of the string doesn't match.

TODO: develop and use a precedence rule to generalise the parser, which
also helps in cleaning up the codebase.

commit-by: Huy Nguyen