Generalize the matching functions
The first one in the chain (match_step_index) returns the index so that
the substring from 0 to that index of the string passed in matches the
regex. The second one (match_step_substr) returns the substring implied
by match_step_index if there's a match. The third (match_first) returns
the first substring that matches the regex. The last (is_match) works
like before.

commit-by: Huy Nguyen